Move tests for implicits to its own file
